move daemonset resource to its own file for clarity
